+++ src/F1Telemetry.F125/Packets/LapPositionsPacket.cs
@@ -4,5 +4,5 @@ public sealed class LapPositionsPacket
 {
     public byte NumLaps { get; set; }
     public byte LapStart { get; set; }
-    public byte[][] PositionForVehicleIdx { get; set; } = [];
+    public int[][] PositionForVehicleIdx { get; set; } = [];
 }
